requirements.txt
openai
requests
markdownify
googlesearch-python
beautifulsoup4
lxml
json5
kaggle
pandas
xlsxwriter
